Platform Collisions
- Changed the one sided mesh collider system with trigger based system.
By default all platform colliders are triggers. The moment player is in
X bounds of the platform and in higher position in Y from the platform,
the platform becomes a rigid collider. This fixes the pops in collision
and unwanted collision response on platform sides when the player tries
to jump on them.